requirements.txt
# Core dependencies
lightrag-hku>=1.3.8
networkx>=3.0
numpy>=1.24
openai>=1.0
pandas>=2.0
scipy>=1.10
aiohttp>=3.8
requests>=2.28
chardet>=5.0

# Dev
pytest>=7.0
